ubbang/backend/requirements.txt
fastapi
sqlalchemy
pydantic
boto3
redis
requests
openai
python-dotenv
passlib[bcrypt]
langchain
langchain-core
langchain-openai
uvicorn
pymysql
httpx
langchain-community
faiss-cpu
uvicorn[standard]
python-jose
pydantic
sympy